Sixth milestone (edited)

+++ job.c
@@ -5,6 +5,13 @@ int jobCount = 0;
 int nextJobId = 1;
 pid_t shell_pgid;
 
+/**
+ * @name cmdJobs
+ * @brief print out jobs that are currently running or stopped.
+ * @param None
+ * @returns void, only print  
+ */
+
 void cmdJobs(){
     for (int i = 0; i < jobCount; i++){
         if (jobs[i].running == true) {
@@ -15,36 +22,48 @@ void cmdJobs(){
         }
     }
 }
-void cmdFg(char* buffer){
-    char* temp = strchr(buffer, '%');
-    memmove(temp, temp+1, strlen(temp)); //Get Job ID
-    int jobId = atoi(temp);
+
+/**
+ * @name cmdFg
+ * @brief Bring background processes to foreground and run
+ * @param buffer argv[1] e.g. %1
+ * @param argc number of arguments
+ * @returns void, only bring
+ */
+
+void cmdFg(char* buffer, int argc){
+
+    if (argc != 2){ // If not what we wanted
+        printf("Example fg command: 'fg %%1'\n");
+        return;
+    }
+
+    memmove(buffer, buffer+1, strlen(buffer)); //Get Job ID
+    int jobId = atoi(buffer);
 
     for (int i = 0; i < jobCount; i++) {
         if (jobs[i].id == jobId) {
             pid_t pid = jobs[i].pid;
 
-            kill(pid, SIGCONT);
+            kill(pid, SIGCONT); // Continue the process
 
-            tcsetpgrp(0, pid);
+            tcsetpgrp(0, pid); // Give terminal to child process
             printf("%s\n", jobs[i].command);
             fflush(stdout);
             int stat;
-            waitpid(pid, &stat, WUNTRACED); 
+            waitpid(pid, &stat, WUNTRACED); // Wait for the child to finish
 
             tcsetpgrp(0, shell_pgid);
 
-            if (WIFSTOPPED(stat)) {
+            if (WIFSTOPPED(stat)) { // If the child is stopped
                 jobs[i].running = false;
                 printf("\n[%d] Stopped\t%s\n", jobs[i].id, jobs[i].command);
-            } else {
-
-                for (int j = i; j < jobCount - 1; j++) {
-                    jobs[j] = jobs[j + 1];
-                }
-                jobCount--;
+            }
+            else { // If not stopped then remove that job
+                removeJob(i); 
             }
 
+            // Update exit code
             if (WIFEXITED(stat)) {
                 prevExitCode = WEXITSTATUS(stat);
             } else {
@@ -55,10 +74,24 @@ void cmdFg(char* buffer){
         }
     }
 }
-void cmdBg(char* buffer) {
-    char* temp = strchr(buffer, '%');
-    memmove(temp, temp+1, strlen(temp)); //Get Job ID
-    int jobId = atoi(temp);
+
+/**
+ * @name cmdBg
+ * @brief Bring stopped processes to run in background
+ * @param buffer argv[1] e.g. %1
+ * @param argc number of arguments
+ * @returns vonid, only brig
+ */
+
+void cmdBg(char* buffer, int argc) {
+
+    if (argc != 2){ // If not what we wanted
+        printf("Example fg command: 'bg %%1'\n");
+        return;
+    }
+
+    memmove(buffer, buffer+1, strlen(buffer)); //Get Job ID
+    int jobId = atoi(buffer);
 
     for (int i = 0; i < jobCount; ++i) {
         if (jobs[i].id == jobId) {
@@ -72,8 +105,16 @@ void cmdBg(char* buffer) {
             return;
         }
     }
+    prevExitCode = 1; // Can't find the given job id
 }
 
+/**
+ * @name createJob
+ * @brief Create a job and put in the necessary details
+ * @param command First argument
+ * @param pid Process id that we are building for
+ * @returns a new Job in jobs list
+ */
 Job createJob(char* command, pid_t pid){
     Job job;
     job.id = nextJobId++;
@@ -83,4 +124,18 @@ Job createJob(char* command, pid_t pid){
     job.done = false;
     jobs[jobCount++] = job;
     return job;
+}
+
+/**
+ * @name removeJob
+ * @brief Removing a job from jobs list
+ * @param index Locate which job to remove
+ * @returns void, just shifting a list
+ */
+void removeJob(int index){
+    for (int j = index; j < jobCount - 1; j++) {
+        jobs[j] = jobs[j + 1];
+    }
+    jobCount--;
+    return;
 }
